added enum support for esi. characterId is an int.

+++ model/ESI/src/main/java/fr/guiguilechat/eveonline/model/esi/modeled/Verify.java
@@ -5,7 +5,7 @@ import fr.guiguilechat.eveonline.model.esi.direct.ESIRawConnection.R_Verify;
 
 /**
  * access to the verify method( not in swagger)
- * 
+ *
  */
 public class Verify {
 
@@ -24,7 +24,7 @@ public class Verify {
 		return verify;
 	}
 
-	public long characterID() {
+	public int characterID() {
 		return verify().CharacterID;
 	}
